+++ CHANGELOG.md
@@ -1,9 +1,23 @@
 # Changelog
 All changes to the project will be documented in this file.
 
+## 1.0.2 - 2023.07.12
+
+### Commands added
+
+- Added `x-move-win-name-screen` to move Windows on a specific screen
+
+### Other
+
+- Implement command callbacks in addition to parsing callbacks
+
 ## 1.0.2 - 2023.07.11
 
-- Added command x-focus-name-nth to focus on the nth instance of the underlying window name
+### Commands added
+
+- Added `command x-focus-name-nth` to focus on the nth instance of the underlying window name
+
+### Other
 - Added the -x option to wmctrl
 - Script can be executed without setting the working direcotry
 - Added this changelog